@dafeda 2024-01-03
Ran with and without trick on a gaussian field with 1 million parameters, 50 observations and 25 realizations. The sigma=10 controls the smoothness of the field.
It runs faster without the trick.

```
    p = 1 mill, m = 50, N = 25, sigma = 10
    With trick:    1.8s, 1.7s, 1.9s (Mean)
    Without trick: 1.3s, 1.2s, 1.3s (Mean)
```

![image](https://github.com/equinor/iterative_ensemble_smoother/assets/45088507/0dd041d9-4274-4410-9165-14129870a905)

I also ran Drogon with and without the trick and it runs much faster without the trick
My conclusion for now is that the trick is faster when the field is very smooth which it does not seem to be in practice (Drogon).
So, I suggest we remove it.